+++ drivers/spirv/src/main/java/uk/ac/manchester/tornado/drivers/spirv/mm/SPIRVIntArrayWrapper.java
@@ -29,11 +29,11 @@ import uk.ac.manchester.tornado.drivers.spirv.SPIRVDeviceContext;
 public class SPIRVIntArrayWrapper extends SPIRVArrayWrapper<int[]> {
 
     public SPIRVIntArrayWrapper(SPIRVDeviceContext device, long batchSize) {
-        super(device, JavaKind.Byte, batchSize);
+        super(device, JavaKind.Int, batchSize);
     }
 
     protected SPIRVIntArrayWrapper(final int[] array, final SPIRVDeviceContext device, long batchSize) {
-        super(array, device, JavaKind.Byte, batchSize);
+        super(array, device, JavaKind.Int, batchSize);
     }
 
     @Override
